Fixup handling of params[:ha] in component_runit_service
params[:ha] can be false, but our existing logic used || as a shortcut
mechanism which didn't work.  Rewrote how we check params[:ha].

Add in extra param :package, which defaults to 'private_chef' and
is for which part of the node hierarchy you go down.  It avoids a lot of
boilerplate setting of params in add-ons