[OC-11575] Don't start services by default in HA topology.
In the HA topology, keepalived is responsible for starting and
stopping services based on its master election.  Without rendering a
down file, runsv would attempt to start services automatically,
leading to a number of race conditions on boot.

The `down` file is a runsv convention that prevents the service from
being started automatically.  It can still be started manually by
sending the start command via sv.