Fix use of undefined variables and add tests
In the pg database and user resources a connection_string variable was
being referenced that was not defined in the correct places. This fixes
that by extracting it out into a method and adds ChefSpec tests for
basic verification of the command output.

Passing the `--password` flag to `createdb` does not actually work how
the code expects. These commits do not fix that, but leave the behavior
as was intended in the original code. Another PR will be needed to fix
that.