Set systemd TasksMax to infinity
Systemd 228 reduced the max number of tasks allowed to 512, which when
exceeded causes solr to fail with the somewhat opaque message
'java.lang.OutOfMemoryError: unable to create new native
thread'. Since the runit process group encompasses all of the chef
services, that isn't too hard to hit.

SLES-12 and Ubuntu 16.04 appear to include this version of Systemd. In
particular, our SLES-12 tester was hitting this during pedant runs,
and all the search tests were imploding in our face.

Thanks to @nsdavidson for the detective work that got to the root cause.

References:
The helpful post that got us on the right track:
https://www.elastic.co/blog/we-are-out-of-memory-systemd-process-limits

The systemd issue created:
https://github.com/systemd/systemd/issues/3211

Documentation on the TasksMax setting:
man systemd.resource-control

Signed-off-by: Mark Anderson <mark@chef.io>