Override runit init template for rabbitmq.
The runit cookbook fixed an old bug which was symlinking the sv binary
to the init.d directory which would not evaluate correctly on
Debian/Ubuntu. This script contains a hard-coded path which breaks the
omnibus-ctl command to start rabbitmq as the path needed to the Erlang
binary is not in the path of the script. This allows us to override the
value for this script and fix the bug.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>