Pin runit cookbook to v4.1.1.
There was a bugfix in the runit cookbook introduced in v4.1.2 which
changed the init.d from a symlink of the sv binary to a init.d script.
This script correctly has a hard-coded `PATH` variable which does not
include the omnibus path, causing the rabbitmq init script to fail via
chef-server-ctl/runit when it is invoked manually. Unfortunately, there
seems to be some scope issues with custom resources in Chef 14 which
are preventing us from overriding the template used by the runit
cookbook without a significant effort in rewriting all the runit usage
in this and the private-chef cookbooks.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>