turn component_runit_service into a custom resource
Initiated to appease ChefSpec 7.3 in omnibus projects that depend on
this cookbook. ChefSpec 7.3+ no longer supports resource definitions and
will error in test runs complaining about a missing params method.

This converts the component_runit_service resource into a custom
resource and adds a ChefSpec spec for it. This aims for backwards
compatibility and does not attempt to improve the resource in any way.

Signed-off-by: Robb Kidd <robb@thekidds.org>