add some actions that delegate to the runit_service
These actions are often called on the runit_service directly by the
cookbooks that use this resource, generally as a notifies action. This
is hard to test for and feels dirty.

So, add some actions on the component_runit_service that are essentially
delegates to the runit_service resource so notifies can be wired up to
a component_runit_service in the dependent cookbooks.

Signed-off-by: Robb Kidd <robb@thekidds.org>